+++ src/model/category.rs
@@ -4,7 +4,7 @@ use super::page::Page;
 use serde::{Deserialize, Serialize};
 /// category object
 /// [category object](https://developer.spotify.com/web-api/get-list-categories/#categoryobject)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct Category {
     pub href: String,
     pub icons: Vec<Image>,
@@ -14,7 +14,7 @@ pub struct Category {
 
 /// Categories wrapped by page object
 /// [get list categories](https://developer.spotify.com/web-api/get-list-categories/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct PageCategory {
     pub categories: Page<Category>,
 }