+++ src/model/image.rs
@@ -2,7 +2,7 @@
 use serde::{Deserialize, Serialize};
 
 /// [image object](https://developer.spotify.com/documentation/web-api/reference/object-model/#image-object)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct Image {
     pub height: Option<u32>,
     pub url: String,