+++ src/model/playing.rs
@@ -7,7 +7,7 @@ use super::track::FullTrack;
 
 /// Playing history
 /// [play history object](https://developer.spotify.com/documentation/web-api/reference/object-model/#play-history-object)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct PlayHistory {
     pub track: FullTrack,
     pub played_at: DateTime<Utc>,