+++ src/model/recommend.rs
@@ -3,13 +3,13 @@ use super::track::SimplifiedTrack;
 use serde::{Deserialize, Serialize};
 
 /// [Recommendations object](https://developer.spotify.com/documentation/web-api/reference/object-model/#recommendations-object)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct Recommendations {
     pub seeds: Vec<RecommendationsSeed>,
     pub tracks: Vec<SimplifiedTrack>,
 }
 /// [Recommendations seed object](https://developer.spotify.com/documentation/web-api/reference/object-model/#recommendations-seed-object)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct RecommendationsSeed {
     #[serde(rename = "afterFilteringSize")]
     pub after_filtering_size: u32,
@@ -23,7 +23,7 @@ pub struct RecommendationsSeed {
     pub _type: RecommendationsSeedType,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub enum RecommendationsSeedType {
     #[serde(rename = "ARTIST")]
     Artist,