+++ src/model/search.rs
@@ -8,39 +8,39 @@ use super::track::FullTrack;
 use serde::{Deserialize, Serialize};
 
 ///[Search item](https://developer.spotify.com/web-api/search-item/);
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchPlaylists {
     pub playlists: Page<SimplifiedPlaylist>,
 }
 ///[Search item](https://developer.spotify.com/web-api/search-item/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchAlbums {
     pub albums: Page<SimplifiedAlbum>,
 }
 
 ///[Search item](https://developer.spotify.com/web-api/search-item/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchArtists {
     pub artists: Page<FullArtist>,
 }
 ///[Search item](https://developer.spotify.com/web-api/search-item/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchTracks {
     pub tracks: Page<FullTrack>,
 }
 
 /// [Search item](https://developer.spotify.com/web-api/search-item/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchShows {
     pub shows: Page<SimplifiedShow>,
 }
 /// [Search item](https://developer.spotify.com/web-api/search-item/)
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SearchEpisodes {
     pub episodes: Page<SimplifiedEpisode>,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub enum SearchResult {
     #[serde(rename = "playlists")]
     Playlists(Page<SimplifiedPlaylist>),